#376: rename LBAF bach to LBAF_app
