+++ docs/pages/input_data.rst
@@ -1,7 +1,7 @@
 Input data
 ==========
 
-`LBAF.py` supports two kind on input data.
+`LBAF_app.py` supports two kind on input data.
 
 LB data
 -------
