+++ docs/pages/usage.rst
@@ -9,7 +9,7 @@ In order to run LBAF from main project directory
 .. code-block:: bash
 
   cd <project-path>
-  python3 src/lbaf/Applications/LBAF.py --configuration=conf.yaml
+  python3 src/lbaf/Applications/LBAF_app.py --configuration=conf.yaml
 
 InformAndtransfer algorithm
 ---------------------------
