+++ scripts/run_tests.sh
@@ -5,5 +5,5 @@ export DISPLAY=:99.0
 /bin/sh /lbaf/venv/bin/activate
 Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
 pip install -r requirements.txt
-python /lbaf/src/Applications/LBAF.py
+python /lbaf/src/Applications/LBAF_app.py
 python /lbaf/tests/test_lbaf.py
